test(ui): split TC-UX-007 per AGENTS.md "one .spec.ts per test case" rule
The first revision of TC-UX-007 bundled the default-kind (quotes) and the ?kind=order (orders) scenarios into a single file with two test() blocks. .ai/qa/AGENTS.md requires one .spec.ts per test case, matching the numbering model used throughout the suite. - TC-UX-007 now carries only the default-kind case (stale id renders RecordNotFoundState with the "Back to quotes" back link). - TC-UX-009 carries the kind=order case (stale id with ?kind=order renders RecordNotFoundState with the "Back to orders" back link). - Spec changelog updated to reference the four files (006/007/008/009). No production-code change; integration discovery still reports four tests.
