+++ packages/core/src/modules/sales/__integration__/TC-UX-007-record-not-found-documents.spec.ts
@@ -2,16 +2,16 @@ import { expect, test } from '@playwright/test'
 import { login } from '@open-mercato/core/helpers/integration/auth'
 
 /**
- * TC-UX-007: RecordNotFoundState renders for a missing sales document detail
+ * TC-UX-007: RecordNotFoundState renders for a missing sales document detail (default kind → quotes)
  * Source: .ai/specs/2026-03-23-unified-record-not-found-ui-state.md (Phase 5 — Integration Coverage)
  *
  * The sales documents detail page is backed by a list API that returns an
  * empty `items` array when the requested id does not exist (not an HTTP 404).
- * This test exercises that branch: a stale/random id should render the shared
- * `RecordNotFoundState` instead of the detail form.
+ * This test exercises that branch with no `kind` query parameter, so the
+ * back-to-list action should point at the quotes list.
  */
-test.describe('TC-UX-007: RecordNotFoundState — sales document detail with stale id', () => {
-  test('renders shared not-found state and back-to-list action for an unknown document', async ({ page }) => {
+test.describe('TC-UX-007: RecordNotFoundState — sales document detail with stale id (default kind)', () => {
+  test('renders shared not-found state and back-to-quotes action for an unknown document', async ({ page }) => {
     await login(page, 'admin')
 
     const staleId = crypto.randomUUID()
@@ -23,19 +23,5 @@ test.describe('TC-UX-007: RecordNotFoundState — sales document detail with sta
     const backLink = page.getByRole('link', { name: /back to quotes/i })
     await expect(backLink).toBeVisible()
     await expect(backLink).toHaveAttribute('href', '/backend/sales/quotes')
-  })
-
-  test('renders order-scoped back link when kind=order is supplied', async ({ page }) => {
-    await login(page, 'admin')
-
-    const staleId = crypto.randomUUID()
-    await page.goto(`/backend/sales/documents/${staleId}?kind=order`, { waitUntil: 'commit' })
-
-    const notFoundLabel = page.getByText('Document not found', { exact: false })
-    await expect(notFoundLabel).toBeVisible({ timeout: 15_000 })
-
-    const backLink = page.getByRole('link', { name: /back to orders/i })
-    await expect(backLink).toBeVisible()
-    await expect(backLink).toHaveAttribute('href', '/backend/sales/orders')
   })
 })
